Switched to CMake
Changes:
* Removal of old configure & makefiles
* Completely switch to CMake

Enhancements:
* Removal of warnings in src/smbase/mysig.cc
* Removal of warnings in src/smbase/mypopen.c
* Removal of +x in src/smbase/malloc_stub.c

Fixes:
* Fixed src/elkhound/examples/scannerless

Cleanup:
* Removal of src/elkhound/triv/DeclExpr*
* Removal of src/elkhound/asfsdf
* Removal of src/elkhound/out
* Removal of src/smbase/.gdbinit

+++ src/smbase/mysig.cc
@@ -164,9 +164,9 @@ int main(int argc, char **argv)
       infiniteRecursion();
     }
 
-    int addr = strtoul(argv[1], NULL /*endp*/, 0 /*radix*/);
+    intptr_t addr = strtoul(argv[1], NULL /*endp*/, 0 /*radix*/);
     printf("about to access 0x%08X ...\n", addr);
-    *((int*)addr) = 0;
+    *((intptr_t*)addr) = 0;
     return 0;     // won't be reached for most values of 'addr'
   }
